chore: delete unnessary area demo
