--- examples/data/basement.json
@@ -1,157 +1,0 @@
-[
-  {
-    "time": "2018-08-28",
-    "UV": 719,
-    "PV": 16238
-  },
-  {
-    "time": "2018-08-29",
-    "UV": 713,
-    "PV": 16813
-  },
-  {
-    "time": "2018-08-30",
-    "UV": 696,
-    "PV": 16633
-  },
-  {
-    "time": "2018-08-31",
-    "UV": 675,
-    "PV": 14378
-  },
-  {
-    "time": "2018-09-01",
-    "UV": 191,
-    "PV": 2506
-  },
-  {
-    "time": "2018-09-02",
-    "UV": 123,
-    "PV": 1922
-  },
-  {
-    "time": "2018-09-03",
-    "UV": 703,
-    "PV": 15742
-  },
-  {
-    "time": "2018-09-04",
-    "UV": 724,
-    "PV": 17403
-  },
-  {
-    "time": "2018-09-05",
-    "UV": 731,
-    "PV": 18656
-  },
-  {
-    "time": "2018-09-06",
-    "UV": 717,
-    "PV": 16581
-  },
-  {
-    "time": "2018-09-07",
-    "UV": 704,
-    "PV": 14078
-  },
-  {
-    "time": "2018-09-08",
-    "UV": 147,
-    "PV": 1911
-  },
-  {
-    "time": "2018-09-09",
-    "UV": 102,
-    "PV": 1226
-  },
-  {
-    "time": "2018-09-10",
-    "UV": 728,
-    "PV": 15862
-  },
-  {
-    "time": "2018-09-11",
-    "UV": 760,
-    "PV": 15709
-  },
-  {
-    "time": "2018-09-12",
-    "UV": 754,
-    "PV": 16754
-  },
-  {
-    "time": "2018-09-13",
-    "UV": 782,
-    "PV": 18184
-  },
-  {
-    "time": "2018-09-14",
-    "UV": 712,
-    "PV": 14982
-  },
-  {
-    "time": "2018-09-15",
-    "UV": 201,
-    "PV": 2966
-  },
-  {
-    "time": "2018-09-16",
-    "UV": 132,
-    "PV": 2494
-  },
-  {
-    "time": "2018-09-17",
-    "UV": 755,
-    "PV": 20528
-  },
-  {
-    "time": "2018-09-18",
-    "UV": 783,
-    "PV": 19201
-  },
-  {
-    "time": "2018-09-19",
-    "UV": 779,
-    "PV": 17008
-  },
-  {
-    "time": "2018-09-20",
-    "UV": 799,
-    "PV": 15903
-  },
-  {
-    "time": "2018-09-21",
-    "UV": 655,
-    "PV": 12557
-  },
-  {
-    "time": "2018-09-22",
-    "UV": 117,
-    "PV": 1160
-  },
-  {
-    "time": "2018-09-23",
-    "UV": 104,
-    "PV": 1476
-  },
-  {
-    "time": "2018-09-24",
-    "UV": 153,
-    "PV": 2485
-  },
-  {
-    "time": "2018-09-25",
-    "UV": 788,
-    "PV": 16616
-  },
-  {
-    "time": "2018-09-26",
-    "UV": 759,
-    "PV": 15439
-  },
-  {
-    "time": "2018-09-27",
-    "UV": 757,
-    "PV": 16786
-  }
-]
